+++ src/api/java/baritone/api/command/datatypes/BlockById.java
@@ -19,27 +19,14 @@ package baritone.api.command.datatypes;
 
 import baritone.api.command.exception.CommandException;
 import baritone.api.command.helpers.TabCompleteHelper;
-
 import net.minecraft.block.Block;
-import net.minecraft.block.properties.IProperty;
 import net.minecraft.init.Blocks;
 import net.minecraft.util.ResourceLocation;
 
-import java.util.Set;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public enum BlockById implements IDatatypeFor<Block> {
     INSTANCE;
-
-    /**
-     * Matches (domain:)?name([(property=value)*])? but the input can be truncated at any position.
-     * domain and name are [a-z0-9_.-]+ and [a-z0-9/_.-]+ because that's what mc 1.13+ accepts.
-     * property and value use the same format as domain.
-     */
-    // Good luck reading this.
-    private static Pattern PATTERN = Pattern.compile("(?:[a-z0-9_.-]+:)?(?:[a-z0-9/_.-]+(?:\\[(?:(?:[a-z0-9_.-]+=[a-z0-9_.-]+,)*(?:[a-z0-9_.-]+(?:=(?:[a-z0-9_.-]+(?:\\])?)?)?)?|\\])?)?)?");
 
     @Override
     public Block get(IDatatypeContext ctx) throws CommandException {
@@ -53,111 +40,14 @@ public enum BlockById implements IDatatypeFor<Block> {
 
     @Override
     public Stream<String> tabComplete(IDatatypeContext ctx) throws CommandException {
-        String arg = ctx.getConsumer().getString();
-
-        if (!PATTERN.matcher(arg).matches()) {
-            // Invalid format; we can't complete this.
-            return Stream.empty();
-        }
-
-        if (arg.endsWith("]")) {
-            // We are already done.
-            return Stream.empty();
-        }
-
-        if (!arg.contains("[")) {
-            // no properties so we are completing the block id
-            return new TabCompleteHelper()
-                    .append(
-                            Block.REGISTRY.getKeys()
-                                    .stream()
-                                    .map(Object::toString)
-                    )
-                    .filterPrefixNamespaced(arg)
-                    .sortAlphabetically()
-                    .stream();
-        }
-
-        // destructuring assignment? Please?
-        String blockId, properties;
-        {
-            String[] parts = splitLast(arg, '[');
-            blockId = parts[0];
-            properties = parts[1];
-        }
-
-        Block block = Block.REGISTRY.getObject(new ResourceLocation(blockId));
-        if (block == null) {
-            // This block doesn't exist so there's no properties to complete.
-            return Stream.empty();
-        }
-
-        String leadingProperties, lastProperty;
-        {
-            String[] parts = splitLast(properties, ',');
-            leadingProperties = parts[0];
-            lastProperty = parts[1];
-        }
-
-        if (!lastProperty.contains("=")) {
-            // The last property-value pair doesn't have a value yet so we are completing its name
-            Set<String> usedProps = Stream.of(leadingProperties.split(","))
-                    .map(pair -> pair.split("=")[0])
-                    .collect(Collectors.toSet());
-
-            String prefix = arg.substring(0, arg.length() - lastProperty.length());
-            return new TabCompleteHelper()
-                    .append(
-                            block.getBlockState()
-                                    .getProperties()
-                                    .stream()
-                                    .map(IProperty::getName)
-                    )
-                    .filter(prop -> !usedProps.contains(prop))
-                    .filterPrefix(lastProperty)
-                    .sortAlphabetically()
-                    .map(prop -> prefix + prop)
-                    .stream();
-        }
-
-        String lastName, lastValue;
-        {
-            String[] parts = splitLast(lastProperty, '=');
-            lastName = parts[0];
-            lastValue = parts[1];
-        }
-
-        // We are completing the value of a property
-        String prefix = arg.substring(0, arg.length() - lastValue.length());
-
-        IProperty<?> property = block.getBlockState().getProperty(lastName);
-        if (property == null) {
-            // The property does not exist so there's no values to complete
-            return Stream.empty();
-        }
-
         return new TabCompleteHelper()
-                .append(getValues(property))
-                .filterPrefix(lastValue)
+                .append(
+                        Block.REGISTRY.getKeys()
+                                .stream()
+                                .map(Object::toString)
+                )
+                .filterPrefixNamespaced(ctx.getConsumer().getString())
                 .sortAlphabetically()
-                .map(val -> prefix + val)
                 .stream();
-    }
-
-    /**
-     * Always returns exactly two strings.
-     * If the separator is not found the FIRST returned string is empty.
-     */
-    private static String[] splitLast(String string, char chr) {
-        int idx = string.lastIndexOf(chr);
-        if (idx == -1) {
-            return new String[]{"", string};
-        }
-        return new String[]{string.substring(0, idx), string.substring(idx + 1)};
-    }
-
-    // this shouldn't need to be a separate method?
-    private static <T extends Comparable<T>> Stream<String> getValues(IProperty<T> property) {
-        return property.getAllowedValues().stream().map(property::getName);
     }
 }
